requirements.txt
torch>=1.9.0
numpy>=1.19.2
PyYAML>=5.4.1
matplotlib>=3.3.4
tensorboard>=2.4.1
pillow>=8.2.0
tqdm>=4.61.0
scikit-learn>=0.24.2
opencv-python>=4.5.2 